Fix subtle but important error in Gustafson-Barsis law.
This law uses a parameter that is the time in parallel execution spent
in inherently serial code. I had it listed as the fraction of inherently
serial code (i.e. serial execution spent in inherently serial code).
I also added functional notation, since that bothered one of the reviewers
(which ended up alerting me to the problem).