Re-add a note about measuring by energy.
Reviewer 3 complained about using the term cost as core-hours since
that might be confused with dollar cost or power cost. I disagree.
Partly because the term cost has precident in previous work. Partly
because there is no reason one couldn't use some other form of cost
(such as electrical power) in the derivations. A mention of the former
is already there. I re-introduced a concluding paragraph describing the
latter.